+++ src/inspect_ai/_display/textual/widgets/samples.py
@@ -31,7 +31,7 @@ from inspect_ai._display.textual.widgets.vscode import conditional_vscode_link
 from inspect_ai._util.file import to_uri
 from inspect_ai._util.format import format_progress_time
 from inspect_ai._util.port_names import get_service_by_port
-from inspect_ai._util.registry import registry_unqualified_name
+from inspect_ai._util.task import task_display_name
 from inspect_ai._util.vscode import EXTENSION_COMMAND_OPEN_SAMPLE, VSCodeCommand
 from inspect_ai.event._tool import ToolEvent
 from inspect_ai.log._samples import ActiveSample
@@ -162,7 +162,7 @@ class SamplesList(OptionList):
             table.add_column(width=20)
             table.add_column(width=11, justify="right")
             table.add_column(width=1)
-            task_name = Text.from_markup(f"{registry_unqualified_name(sample.task)}")
+            task_name = Text.from_markup(f"{task_display_name(sample.task)}")
             task_name.truncate(18, overflow="ellipsis", pad=True)
             task_time = Text.from_markup(f"{format_progress_time(sample.running_time)}")
             table.add_row(task_name, task_time, " ")
@@ -328,7 +328,7 @@ class SampleInfo(Vertical):
 
             # update UI
             self.display = True
-            title = f"{registry_unqualified_name(sample.task)} (id: {sample.sample.id}, epoch {sample.epoch}): {sample.model}"
+            title = f"{task_display_name(sample.task)} (id: {sample.sample.id}, epoch {sample.epoch}): {sample.model}"
             self.query_one(Collapsible).title = title
             sandboxes = self.query_one(SandboxesView)
             await sandboxes.sync_sample(sample)