+++ src/inspect_ai/_view/www/src/components/PopOver.tsx
@@ -101,18 +101,21 @@ export const PopOver: React.FC<PopOverProps> = ({
     if (!isOpen || hoverDelay <= 0) {
       setShouldShowPopover(isOpen);
       const listener = (event: MouseEvent) => {
-        // Only close if clicking outside the popover content
+        // Only close if clicking outside the popover content,
+        // prevent default actions "in the background" when using click to close the popover
         if (
           popperRef.current &&
           !popperRef.current.contains(event.target as Node)
         ) {
+          event.preventDefault();
+          event.stopPropagation();
           setIsOpen(false);
         }
       };
-      document.addEventListener("mousedown", listener);
+      document.addEventListener("click", listener, true);
 
       return () => {
-        document.removeEventListener("mousedown", listener);
+        document.removeEventListener("click", listener, true);
       };
     }